CPUCache: silence CodeQL warnings

+++ src/modules/cpucache/cpucache.c
@@ -44,7 +44,7 @@ static void printCPUCacheNormal(const FFCPUCacheResult* result, FFCPUCacheOption
             ffParseSize(src->size, &buffer);
             ffStrbufAppendF(&buffer, " (%c)", typeStr);
 
-            sum += src->size * src->num;
+            sum += (uint64_t) src->size * src->num;
         }
 
         if(options->moduleArgs.outputFormat.length == 0)
@@ -74,7 +74,7 @@ static void printCPUCacheCompact(const FFCPUCacheResult* result, FFCPUCacheOptio
             ffStrbufAppendS(&buffer, ", ");
         uint32_t value = 0;
         FF_LIST_FOR_EACH(FFCPUCache, src, result->caches[i])
-            value += src->size * src->num;
+            value += (uint64_t) src->size * src->num;
         ffParseSize(value, &buffer);
         ffStrbufAppendF(&buffer, " (L%u)", i + 1);
         sum += value;
